Add skill_index.py
